// src/generated/zod.gen.ts
export const zListLinksData = z.object({
    body: z.optional(z.never()),
    path: z.optional(z.never()),
    query: z.object({
        domain_id: z.int().gte(1),
        limit: z.optional(z.int().gte(1).lte(150)),
        idString: z.optional(z.string()),
        createdAt: z.optional(z.string()),
        beforeDate: z.optional(z.iso.datetime()),
        afterDate: z.optional(z.iso.datetime()),
        dateSortOrder: z.optional(z.enum(['asc', 'desc'])),
        pageToken: z.optional(z.string()),
        folderId: z.optional(z.string())
    }),
    headers: z.optional(z.object({
        'x-vercel-id': z.optional(z.string())
    }))
});
// ...
export const zGetLinkData = z.object({
    body: z.optional(z.never()),
    path: z.object({
        linkId: z.string().regex(/lnk_\w+|\d+/)
    }),
    query: z.optional(z.object({
        domainId: z.optional(z.string())
    }))
});
// ...
export const zCreateLinkData = z.object({
    body: z.optional(z.object({
        originalURL: z.string(),
        cloaking: z.optional(z.boolean()),
        password: z.optional(z.string()),
        redirectType: z.optional(z.union([
            z.literal(301),
            z.literal(302),
            z.literal(307),
            z.literal(308)
        ])),
        expiresAt: z.optional(z.union([
            z.int(),
            z.string()
        ])),
        expiredURL: z.optional(z.string()),
        title: z.optional(z.string()),
        tags: z.optional(z.array(z.string())),
        utmSource: z.optional(z.string()),
        utmMedium: z.optional(z.string()),
        utmCampaign: z.optional(z.string()),
        utmTerm: z.optional(z.string()),
        utmContent: z.optional(z.string()),
        ttl: z.optional(z.union([
            z.string(),
            z.int()
        ])),
        path: z.optional(z.string()),
        androidURL: z.optional(z.string()),
        iphoneURL: z.optional(z.string()),
        createdAt: z.optional(z.union([
            z.int(),
            z.string()
        ])),
        clicksLimit: z.optional(z.int().gte(1)),
        passwordContact: z.optional(z.boolean()),
        skipQS: z.optional(z.boolean()).default(false),
        archived: z.optional(z.boolean()).default(false),
        splitURL: z.optional(z.string()),
        splitPercent: z.optional(z.int().gte(1).lte(100)),
        integrationAdroll: z.optional(z.string()),
        integrationFB: z.optional(z.string()),
        integrationTT: z.optional(z.string()),
        integrationGA: z.optional(z.string()),
        integrationGTM: z.optional(z.string())
    }).and(z.object({
        domain: z.string(),
        allowDuplicates: z.optional(z.boolean()).default(false),
        folderId: z.optional(z.string()),
        skipQS: z.optional(z.boolean())
    }))),
    path: z.optional(z.never()),
    query: z.optional(z.never())
});

/**
 * Default Response
 */
export const zCreateLinkResponse = z.object({
    originalURL: z.string(),
    cloaking: z.optional(z.boolean()),
    password: z.optional(z.string()),
    expiresAt: z.optional(z.union([
        z.int(),
        z.string()
    ])),
    expiredURL: z.optional(z.string()),
    title: z.optional(z.string()),
    tags: z.optional(z.array(z.string())),
    utmSource: z.optional(z.string()),
    utmMedium: z.optional(z.string()),
    utmCampaign: z.optional(z.string()),
    utmTerm: z.optional(z.string()),
    utmContent: z.optional(z.string()),
    ttl: z.optional(z.union([
        z.string(),
        z.int()
    ])),
    path: z.optional(z.string()),
    androidURL: z.optional(z.string()),
    iphoneURL: z.optional(z.string()),
    createdAt: z.optional(z.union([
        z.int(),
        z.string()
    ])),
    clicksLimit: z.optional(z.int().gte(1)),
    passwordContact: z.optional(z.boolean()),
    skipQS: z.optional(z.boolean()).default(false),
    archived: z.optional(z.boolean()).default(false),
    splitURL: z.optional(z.string()),
    splitPercent: z.optional(z.int().gte(1).lte(100)),
    integrationAdroll: z.optional(z.string()),
    integrationFB: z.optional(z.string()),
    integrationTT: z.optional(z.string()),
    integrationGA: z.optional(z.string()),
    integrationGTM: z.optional(z.string())
}).and(z.object({
    idString: z.string(),
    id: z.string(),
    shortURL: z.string(),
    secureShortURL: z.string(),
    path: z.string(),
    redirectType: z.optional(z.enum([
        '301',
        '302',
        '307',
        '308'
    ])),
    FolderId: z.optional(z.string()),
    DomainId: z.optional(z.int()),
    OwnerId: z.optional(z.int()),
    hasPassword: z.optional(z.boolean()).default(false),
    source: z.optional(z.enum([
        'website',
        'api',
        'public',
        'spreadsheets',
        'slack',
        'telegram',
        ''
    ])),
    User: z.optional(z.object({
        id: z.int(),
        name: z.string(),
        email: z.string(),
        photoURL: z.string()
    }))
})).and(z.object({
    success: z.optional(z.boolean()).default(true),
    duplicate: z.optional(z.boolean()).default(false)
}));